reorganize linker integration tests
