+++ Rakefile
@@ -12,7 +12,7 @@ def supported_ruby_box? = RUBY_VERSION >= "4.0.0" && ENV["RUBY_BOX"] == "1" && d
 Rake::TestTask.new do |t|
   t.test_files = FileList['test/**/*_test.rb']
   if supported_ruby_box?
-    t.test_files = FileList['test/caotral/linker/fiddle_test.rb']
+    t.test_files = FileList['test/caotral/linker/integration/fiddle_test.rb']
   end
 end
