+++ test/caotral/linker/integration/fiddle_test.rb
@@ -1,4 +1,4 @@
-require_relative "../../test_suite"
+require_relative "../../../test_suite"
 
 class Caotral::Linker::FiddleMethodTest < Test::Unit::TestCase
   include TestProcessHelper
