+++ test/caotral/linker/integration/gotpc_call_test.rb
@@ -1,4 +1,4 @@
-require_relative "../../test_suite"
+require_relative "../../../test_suite"
 
 class Caotral::Linker::GOTPCRelTest < Test::Unit::TestCase
   include TestProcessHelper
