+++ test/caotral/linker/integration/needed_test.rb
@@ -1,4 +1,4 @@
-require_relative "../../test_suite"
+require_relative "../../../test_suite"
 
 class LinkerNeededTest < Test::Unit::TestCase
   include TestProcessHelper
