+++ test/caotral/linker/integration/pie_object_test.rb
@@ -1,4 +1,4 @@
-require_relative "../../test_suite"
+require_relative "../../../test_suite"
 
 class Caotral::Linker::PIEObjectLinkingTest < Test::Unit::TestCase
   include TestProcessHelper
