+++ test/caotral/linker/integration/symbol_remap_test.rb
@@ -1,4 +1,4 @@
-require_relative "../../test_suite"
+require_relative "../../../test_suite"
 
 class Caotral::Linker::SymbolRemapTest < Test::Unit::TestCase
   include TestProcessHelper
